+++ minify.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+yuglify djangojs/static/js/djangojs/django.js