+++ djangojs/__init__.py
@@ -7,5 +7,5 @@ __version__ = '0.7.3.dev'
 __description__ = "Django JS Tools"
 
 #: Packaged jQuery version
-JQUERY_DEFAULT_VERSION = '1.9.1'
+JQUERY_DEFAULT_VERSION = '2.0.0'
 JQUERY_MIGRATE_VERSION = '1.1.1'